pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator==2.1.0.post1",
    "flask-login==0.6.3",
    "flask==3.0.0",
    "flask-sqlalchemy==3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary==2.9.9",
    "flask-wtf==1.2.1",
    "wtforms==3.1.1",
    "werkzeug==3.0.1",
    "sqlalchemy==2.0.23",
    "twilio>=9.4.6",
    "trafilatura>=2.0.0",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
]
